Added new methods to the utility Strings class.
The following methods were added:
IsNullOrEmpty() - Checks to see if a value convertable to string ends up
as null or empty.
IsNotNullOrEmpty() - Checks to see if the string is not null or empty
IsWhitespaces() - Checks to see if the string is all whitespaces.
IsNotWhitespaces() - Checks to see if the string is not all whitespaces.
Format() - Replaces occurences of {n} with the nth parameter supplied.
First parameter after the format is considered to be parameter 0.